requirements.txt
beautifulsoup4==4.6.0
boto==2.45.0
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.1
dj-static==0.0.6
Django==2.0.5
django-appconf==1.0.2
django-compressor==2.1
django-modelcluster==4.1
django-storages==1.5.1
django-taggit==0.22.2
django-toolbelt==0.0.1
django-treebeard==4.2.0
djangorestframework==3.8.2
draftjs-exporter==2.1.1
gunicorn==19.6.0
html5lib==0.999999999
idna==2.6
l18n==2016.6.4
Markdown==2.6.6
Pillow==5.0.0
pkg-resources==0.0.0
psycopg2==2.6.2
pytz==2018.3
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
six==1.11.0
static3==0.7.0
Unidecode==0.4.21
urllib3==1.22
wagtail==2.1
wagtailmenus==2.9
webencodings==0.5.1
whitenoise==3.2.2
Willow==1.1
